requirements.txt
flask
flask-cors
ultralytics
pillow
numpy
opencv-python-headless
